refactor: move lib/utils to app/utils
